Refactor package structure for CSV-based PGDS
#395
