+++ spark-cypher-examples/src/main/scala/org/opencypher/spark/examples/CustomDataFrameInputExample.scala
@@ -27,7 +27,7 @@
 package org.opencypher.spark.examples
 
 import org.apache.spark.sql.types._
-import org.apache.spark.sql.{DataFrame, Row, SparkSession}
+import org.apache.spark.sql.{DataFrame, Row}
 import org.opencypher.okapi.api.io.conversion.{NodeMapping, RelationshipMapping}
 import org.opencypher.spark.api.CAPSSession
 import org.opencypher.spark.api.io.{CAPSNodeTable, CAPSRelationshipTable}
