+++ spark-cypher-examples/src/main/scala/org/opencypher/spark/examples/CypherSQLRoundtripExample.scala
@@ -28,7 +28,7 @@ package org.opencypher.spark.examples
 
 import org.opencypher.okapi.api.graph.Namespace
 import org.opencypher.spark.api.CAPSSession
-import org.opencypher.spark.api.io.file.FileCsvGraphDataSource
+import org.opencypher.spark.api.io.csv.file.FileCsvGraphDataSource
 import org.opencypher.spark.impl.CAPSConverters._
 
 /**
