+++ spark-cypher-examples/src/main/scala/org/opencypher/spark/examples/Neo4jWorkflow.scala
@@ -28,7 +28,7 @@ package org.opencypher.spark.examples
 
 import org.opencypher.okapi.api.graph.{Namespace, QualifiedGraphName}
 import org.opencypher.spark.api.CAPSSession
-import org.opencypher.spark.api.io.file.FileCsvGraphDataSource
+import org.opencypher.spark.api.io.csv.file.FileCsvGraphDataSource
 import org.opencypher.spark.api.io.neo4j.CommunityNeo4jGraphDataSource
 import org.opencypher.spark.api.io.neo4j.CommunityNeo4jGraphDataSource._
 import org.opencypher.spark.examples.Neo4jHelpers._
